server/requirements/base.txt
Django==4.2.11
djangorestframework==3.15.2

# Auth
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.9.0

# Async & Caching
celery==5.4.0
redis==5.0.4

# Env & Security
python-dotenv==1.0.1

# DB
psycopg[binary]==3.1.19

# HTTP
httpx==0.27.0
